+++ lib/cargo_base.dart
@@ -6,3 +6,4 @@ part 'src/cargo.dart';
 part 'src/shared/memory_impl.dart';
 part 'src/shared/cargo_modes.dart';
 part 'src/shared/cargo_dispatch.dart';
+part 'src/shared/cargo_utils.dart';